requirements.txt
fastapi>=0.115.4
uvicorn>=0.32.0
pydantic>=2.9.2
python-dotenv>=1.0.1
google-cloud-aiplatform>=1.70.0
google-cloud-storage>=2.18.2
langchain-google-vertexai>=2.0.0
langchain-chroma>=0.1.4
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
pypdf>=5.1.0
chromadb>=0.5.18